profile_setting

+++ app/Http/Controllers/frontend/Admin.php
@@ -35,14 +35,14 @@ class Admin extends Controller
 
      public function show_user(Request $request)
      {
-       // if($request->session()->has('u_session')){
+       if($request->session()->has('u_session')){
        $user_get=Register::all();
        // dd($user_get);
        return view('admin.admin_account.user',compact('user_get'));
-     // }else {
+     }else {
 
-       return redirect('/accounts/login');
-     // }
+       return redirect('/login');
+     }
      }
 
 
@@ -53,7 +53,7 @@ class Admin extends Controller
        return view('admin.admin_account.categories',compact('user_get'));
      }else {
 
-       return redirect('/accounts/login');
+       return redirect('/login');
      }
    }
 
@@ -65,7 +65,7 @@ class Admin extends Controller
        return view('admin.admin_account.table');
      }else {
 
-       return redirect('/accounts/login');
+       return redirect('/login');
      }
    }
 
@@ -85,19 +85,20 @@ class Admin extends Controller
      public function admin_dashboard_route(Request $request)
      {
        if($request->session()->has('u_session')){
-       $user_get=DhrUser::get()->count();
-       $active_user=DB::table('dhr_users')->where('active_status','1')->count();
-       $social_users=DB::table('dhr_users')->select('country',DB::raw('COUNT(userId) as count'))->groupBy('country')->get();
+       // $user_get=DhrUser::get()->count();
+       // $active_user=DB::table('dhr_users')->where('active_status','1')->count();
+       // $social_users=DB::table('dhr_users')->select('country',DB::raw('COUNT(userId) as count'))->groupBy('country')->get();
        // dd(count($social_users));
-       $social_users_data=DB::table('dhr_users')->get();
-       $social_users1=count($social_users_data);
+       // $social_users_data=DB::table('registers')->get();
+       // $social_users1=count($social_users_data);
 
 
        // dd($social_users1);
-       return view('admin.admin_account.dashboard',compact('user_get','active_user', 'social_users','social_users1'));
+       // return view('admin.admin_account.dashboard',compact('user_get','active_user', 'social_users','social_users1'));
+       return view('admin.admin_account.dashboard');
      }else {
 
-       return redirect('/accounts/login');
+       return redirect('/login');
      }
    }
 
@@ -108,21 +109,21 @@ class Admin extends Controller
        return view('admin.admin_account.add_category');
      }else {
 
-       return redirect('/accounts/login');
+       return redirect('/login');
      }
    }
 
      public function admin_edit_route(Request $request, $id)
      {
 
-       // if($request->session()->has('u_session')){
+       if($request->session()->has('u_session')){
          // $userinfo= $request->session()->get('u_session')->userId;
          // dd($userinfo);
          $user_get=DB::table('registers')->where('id',$id)->first();
          return view('admin.admin_account.editUser',compact('user_get'));
-       // }else {
-       //   return redirect('/accounts/login');
-       // }
+       }else {
+         return redirect('/login');
+       }
 
      }
 
@@ -136,7 +137,7 @@ class Admin extends Controller
          $user_get=DB::table('skills')->where('skill_id',$id)->first();
          return view('admin.admin_account.editCategory',compact('user_get'));
        }else {
-         return redirect('/accounts/login');
+         return redirect('/login');
        }
 
      }
@@ -146,12 +147,12 @@ class Admin extends Controller
      public function admin_create_route(Request $request)
      {
 
-       // if($request->session()->has('u_session')){
+       if($request->session()->has('u_session')){
          // $userinfo= $request->session()->get('u_session')->userId;
          return view('admin.admin_account.create_user');
-       // }else {
-       //   return redirect('/accounts/login');
-       // }
+       }else {
+         return redirect('/login');
+       }
 
      }
 
@@ -311,15 +312,15 @@ class Admin extends Controller
 
      public function admin_delete_user(Request $request, $id)
      {
-        // if($request->session()->has('u_session')){
+        if($request->session()->has('u_session')){
          // dd($id);
          $user_get=DB::table('registers')->where('id',$id)->delete();
         // dd($user_get);
         echo $user_get;
-      // }else {
-      //
-      //     return redirect('/accounts/login');
-      //   }
+      }else {
+
+          return redirect('/login');
+        }
        }
      public function admin_delete_category(Request $request, $id)
      {
@@ -330,7 +331,7 @@ class Admin extends Controller
         echo $user_get;
       }else {
 
-          return redirect('/accounts/login');
+          return redirect('/login');
         }
        }
